# Bytes -> Hex conversion update