@PunkPun 2023-09-18
👋 

it seems that this solution is incompatible with the mono build

> OpenRA.Game/CryptoUtil.cs(283,40): error CS0117: 'MemoryMarshal' does not contain a definition for 'GetArrayDataReference'