@RoosterDragon 2023-09-18
You can guard with `#if NET6_0_OR_GREATER` and keep the previous implementation around for mono. Alternatively, implementing the loop without any unsafe might sidestep the problem and still be faster.

How does this method compare to [Convert.ToHexString](https://learn.microsoft.com/en-us/dotnet/api/system.convert.tohexstring?view=net-7.0) (which would need `#if NET5_0_OR_GREATER`) followed by a lowercase?