@rudzen 2023-09-18
> You can guard with `#if NET6_0_OR_GREATER` and keep the previous implementation around for mono. Alternatively, implementing the loop without any unsafe might sidestep the problem and still be faster.
> 
> How does this method compare to [Convert.ToHexString](https://learn.microsoft.com/en-us/dotnet/api/system.convert.tohexstring?view=net-7.0) (which would need `#if NET5_0_OR_GREATER`) followed by a lowercase?

Added an alternate implementation for mono using `#if NET5_0_OR_GREATER`, hopefully that should work.
I tried out some other variations, including the Convert.ToHexString() and also BitConverter.ToString() but they seem pretty cumbersome in comparison because of the lowercase and/or '-' removal.

Did not know I could bump to the `#if NET6_0_OR_GREATER`. Will keep that in mind :)