@PunkPun 2023-09-19
The only other place where we use `x2` format is in color code. But it's not hotly used as well