@PunkPun 2023-09-20
> OpenRA/OpenRA.Game/CryptoUtil.cs(269,19): error CS0117: 'Convert' does not contain a definition for 'ToHexString' 

it seems it's incompatible with mono 😩