@rudzen 2023-09-20
> > OpenRA/OpenRA.Game/CryptoUtil.cs(269,19): error CS0117: 'Convert' does not contain a definition for 'ToHexString'
> 
> it seems it's incompatible with mono 😩

It would seem mono doesn't like me that much 😢.
Updated it anyways with a simple none-pre processor version that should work just fine.